Upgrade to latest MSP430-GCC distributed by TI
This improves performance quite a bit.

In  addition  to command  paths  and  flags,  the  following had  to  be
adjusted:

- $r1 does not  seem to exist anymore: operations involving  $base = $r1
  failed  saying  "Argument to  arithmetic  operation  not a  number  or
  boolean."

- The  newer gcc  warns  that short  ints get  promoted  when passed  as
  varargs; the printf function was fixed accordingly.

+++ common/msp/src/printf.c
@@ -488,7 +488,8 @@ int printf (const char *format, ...)
 							else
 #endif /* __MSP430LIBC_PRINTF_INT20__ */
 							{
-								number.i16 = va_arg (args, int16_t);
+                                /* int16_t promoted to int when passed as variable argument */
+								number.i16 = va_arg (args, int);
 								is_zero = (number.i16 == 0);
 								is_negative = (number.i16 < 0);
 							}
